Working towards BMS

+++ embedded/include/can.h
@@ -12,7 +12,7 @@
 CAN_HandleTypeDef hcan;
 
 char* can_read(CAN_HandleTypeDef *hcan);
-void can_send(uint16_t can_id, size_t length, uint8_t* data, CAN_HandleTypeDef *hcan);
+void can_send(uint32_t can_id, size_t length, uint8_t* data, CAN_HandleTypeDef *hcan);
 
 
 uint8_t * can_send_obd2(uint16_t can_id, size_t length, uint8_t mode,uint16_t pid, CAN_HandleTypeDef *hcan);
